Refactor telemetry processing by removing obsolete log batching classes and updating SentryOptions to utilize a telemetry processor. This change enhances the logging mechanism and streamlines the telemetry handling process. Update tests to reflect the new telemetry processor integration.
