+++ packages/dart/lib/src/sentry_options.dart
@@ -16,8 +16,6 @@ import 'telemetry_processing/noop_telemetry_processor.dart';
 import 'telemetry_processing/telemetry_processor.dart';
 import 'transport/noop_transport.dart';
 import 'version.dart';
-import 'sentry_log_batcher.dart';
-import 'noop_log_batcher.dart';
 import 'dart:developer' as developer;
 
 // TODO: shutdownTimeout, flushTimeoutMillis
@@ -541,9 +539,6 @@ class SentryOptions {
   bool includeModuleInStackTrace = false;
 
   late final SentryLogger logger = SentryLogger(clock);
-
-  @internal
-  SentryLogBatcher logBatcher = NoopLogBatcher();
 
   @internal
   TelemetryProcessor telemetryProcessor = NoOpTelemetryProcessor();
